+++ packages/analytics/analytics-geo-map/src/utils/index.ts
@@ -1,11 +1,1 @@
-export type DebounceFunction = (...args: any[]) => void
-
-export function debounce(fn: DebounceFunction, delay: number) {
-  let timeoutId: number
-  return (...args: any[]) => {
-    clearTimeout(timeoutId)
-    timeoutId = window.setTimeout(() => {
-      fn(...args)
-    }, delay)
-  }
-}
+export * from './explore-to-country-metrics'